Move the delete lot button to the lots page.

+++ controllers/lots.js
@@ -1,7 +1,9 @@
 Auction.LotsController = Ember.ArrayController.extend ({
   actions: {
     deleteLot: function(lot) {
-      lot.destroyRecord();
+      if (confirm('Are you sure you want to delete this beautiful lot?')) {
+        lot.destroyRecord();
+      }
     }
   }
 })
